move shared data-fetching code into useProtocols hook
